requirements.txt
Django
django-cors-headers
djangorestframework
pandas
